+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 ## mm/dd/2017
 
 1. [Common](#common)
+    2. [](#bugfix)
+        - Accessibility: Menu / Offcanvas: Offcanvas and hamburger toggle menu are now ARIA compatible (#1891)
     3. [](#bugfix)
         - Menu: Disable Dropdowns still Shows Indicator (#2031)
         - Menu: Fixed frontend menu 'Extended' option, resetting heights of wrapping containers not necessarily related to the menu itself (#2025)